requirements.txt
asgiref==3.8.1
boto3==1.34.149
botocore==1.34.149
chardet==3.0.4
diff-match-patch==20241021
dj-database-url==0.5.0
Django==3.2.25
django-allauth==0.63.6
django-countries==7.2.1
django-crispy-forms==1.14.0
django-import-export==3.0.1
django-storages==1.14.4
django-summernote==0.8.20.0
docutils==0.15.2
et_xmlfile==2.0.0
gunicorn==20.1.0
idna==2.8
jmespath==1.0.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
pillow==10.4.0
psycopg2==2.9.9
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==2.0.0
s3transfer==0.10.2
sqlparse==0.5.1
stripe==10.4.0
tablib==3.8.0
xlrd==2.0.1
xlwt==1.3.0
